kubernetes-feature: реализован функционал наличных

+++ common/src/main/java/com/bank/dto/cash/BalanceDto.java
@@ -3,9 +3,11 @@ package com.bank.dto.cash;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class BalanceDto {
     private Long id;
-    private Long balance;
+    private BigDecimal balance;
 }
